Split the proxy along the seams it already had
proxy.js was one file deciding where a request was going, what to change about it,
how to send it, and how to stand in front of youtube.com while doing so. Four kinds
of request arrive and they look almost alike — one forwarded by a client using us as
an HTTP proxy, one wrapped in the CORS bypass prefix, one that came in over our own
MITM TLS believing it was talking to youtube.com, and a plain one from the page — so
route.js does nothing but tell them apart.

rewrites.js is text in, text out: it reaches no network, holds no state and knows
nothing about an express request, which is what makes it the part that can be
exercised without a set. Whether a rewrite applies stays with knobs.js; this only
knows how.

sending.js keeps the retries, which are worth writing down because neither failure is
the server's fault: a pooled socket that died while idle is handed out anyway and the
request dies on it, and newer Node retries that itself while the Node inside a Tizen
service is old enough that nothing does.

The CA moves out too. It cannot be shipped — one baked into a public release would
put its private key in every download — so each set makes its own on first run and
installs it into the container's trust store.
