+++ service/lib/cobalt.js
@@ -1,87 +1,31 @@
 'use strict';
 
-// Everything the container route needs that a package cannot carry, derived on the set: a
-// writable copy of Cobalt's content directory, and the certificate authority that sits in it.
-// One CA baked into a public release would put its private key in every download.
+// Bringing the container up, and keeping it reachable.
+//
+// Everything a package cannot carry is derived on the set: the trust material is cobaltCa.js, the
+// content directory is cobaltContent.js, and what this package declares about itself is
+// cobaltConfig.js. What is left here is the sequencing — checking the container can reach us,
+// staging and trusting in the right order, and launching it when the platform will not.
 
-const fs = require('fs');
 const os = require('os');
 const path = require('path');
 const dns = require('dns');
 
 const x509 = require('./x509.js');
 const postmortem = require('./postmortem.js');
+const { CONTAINER, appId, configuredContent, container, switches } = require('./cobaltConfig.js');
+const { MITM_DIR, existingMaterial, installCa, issue, stillGood } = require('./cobaltCa.js');
+const { STOCK, cobaltIsInstalledHere, stageOrFail } = require('./cobaltContent.js');
 
-const STOCK = '/usr/apps/com.samsung.tv.cobalt/content/app/cobalt/content';
-
-// The nativeID this package claims, which is also the app id the launched container runs under.
-const CONTAINER = 'com.samsung.tv.cobalt-yt';
-
-const SHARE = process.env.TUBE_SHARE || '/home/owner/share/tube';
-const MITM_DIR = process.env.TUBE_MITM_DIR || path.join(SHARE, 'mitm');
-
-const CONFIG = path.join(__dirname, '..', '..', 'config.xml');
-
-// googlevideo is deliberately absent: media is a blind tunnel, and standing in front of it buys
-// nothing but latency.
-const HOSTS = [
-    'youtube.com', '*.youtube.com',
-    'google.com', '*.google.com',
-    'googleapis.com', '*.googleapis.com',
-    'gstatic.com', '*.gstatic.com',
-    'ggpht.com', '*.ggpht.com'
-];
-
-const REISSUE_WITHIN = 30 * 86400000;
 const RELAUNCH_QUIET = 20000;
 const CLAIM_WITHIN = 10000;
 const LOOKUP_QUIET = 5000;
 const KILL_SETTLE = 1200;
 
-// OpenSSL steps <hash>.0, .1, .2 … when a name collides.
-const SLOTS = 8;
-
 const note = (what, detail) => postmortem.note('cobalt', `${what}: ${postmortem.describe(detail)}`);
 
-const state = { config: undefined, prepared: null, preparing: false, lastWake: 0, waiting: [] };
+const state = { prepared: null, preparing: false, lastWake: 0, waiting: [] };
 const proxied = { context: null, at: 0, lookedAt: 0 };
-
-const config = () => {
-    if (state.config === undefined) {
-        try {
-            state.config = fs.readFileSync(CONFIG, 'utf8');
-        } catch (e) {
-            state.config = null;
-        }
-    }
-
-    return state.config;
-};
-
-// Our own switches, so the staging lands where --content says rather than by convention.
-const switches = () => {
-    const found = /native\.userdata"\s+value="([^"]*)"/.exec(config() || '');
-    return found ? found[1].replace(/&quot;/g, '"') : null;
-};
-
-const configuredContent = () => {
-    if (process.env.TUBE_COBALT_CONTENT) return process.env.TUBE_COBALT_CONTENT;
-
-    const found = /--content=(\S+)/.exec(switches() || '');
-    return found ? found[1] : null;
-};
-
-// The container slot this package claims, or null when it claims none.
-const container = () => {
-    if (config() === null) return null;
-
-    return /--content=|--base_url=/.test(switches() || '') ? CONTAINER : null;
-};
-
-const appId = () => {
-    const found = /<tizen:application\s+id="([^"]+)"/.exec(config() || '');
-    return found ? found[1] : null;
-};
 
 const addresses = () => {
     const interfaces = os.networkInterfaces();
@@ -171,127 +115,6 @@ const wake = () => {
         note('woken', `the container is not up; launching ${me}`);
         return launch(me);
     }, () => {});
-};
-
-// Same size is enough to call a file done: a firmware update replaces the whole directory.
-const copyInto = (from, to) => {
-    fs.mkdirSync(to, { recursive: true });
-
-    return fs.readdirSync(from).reduce((copied, entry) => {
-        const source = path.join(from, entry);
-        const target = path.join(to, entry);
-        const info = fs.statSync(source);
-
-        if (info.isDirectory()) return copied + copyInto(source, target);
-
-        try {
-            if (fs.statSync(target).size === info.size) return copied;
-        } catch (e) { /* absent, so copy it */ }
-
-        fs.writeFileSync(target, fs.readFileSync(source));
-        return copied + 1;
-    }, 0);
-};
-
-const read = (file) => fs.readFileSync(path.join(MITM_DIR, file), 'utf8');
-
-const existingMaterial = () => {
-    try {
-        return { ca: JSON.parse(read('ca.json')), key: read('leaf.key'), chain: read('leaf-chain.crt') };
-    } catch (e) {
-        return null;
-    }
-};
-
-// Reissue well before the cliff, and whenever the set of names has changed.
-const stillGood = (material) => {
-    try {
-        const cert = new (require('crypto').X509Certificate)(material.chain);
-        const left = new Date(cert.validTo).getTime() - Date.now();
-        const names = (cert.subjectAltName || '').replace(/DNS:/g, '').split(', ').sort().join();
-
-        return left > REISSUE_WITHIN && names === HOSTS.slice().sort().join();
-    } catch (e) {
-        // No X509Certificate before Node 15. Trusting what is on disk beats refusing to run.
-        return true;
-    }
-};
-
-const issue = (done) => {
-    const commonName = `Tube Local CA (${os.hostname()})`;
-
-    x509.createCa(commonName, (error, ca) => {
-        if (error) return done(error);
-
-        return x509.createLeaf(ca, 'www.youtube.com', HOSTS, (leafError, leaf) => {
-            if (leafError) return done(leafError);
-
-            try {
-                fs.mkdirSync(MITM_DIR, { recursive: true });
-                fs.writeFileSync(path.join(MITM_DIR, 'ca.crt'), ca.cert);
-                fs.writeFileSync(path.join(MITM_DIR, 'ca.json'), JSON.stringify({ commonName, cert: ca.cert }));
-                fs.writeFileSync(path.join(MITM_DIR, 'leaf-chain.crt'), leaf.chain);
-                fs.writeFileSync(path.join(MITM_DIR, 'leaf.key'), leaf.key, { mode: 0o600 });
-            } catch (writeError) {
-                return done(writeError);
-            }
-
-            note('issued', commonName);
-            return done(null, { ca: { commonName, cert: ca.cert }, key: leaf.key, chain: leaf.chain });
-        });
-    });
-};
-
-// The store is an OpenSSL hashed directory, so the file name is the lookup: <subject_hash>.0, and
-// again under -subject_hash_old because which is used depends on how the verifier was built. A
-// name off by a byte is indistinguishable from an absent CA.
-const installCa = (certs, ca) => {
-    fs.mkdirSync(certs, { recursive: true });
-
-    const contentsOf = (file) => {
-        try { return fs.readFileSync(file, 'utf8'); } catch (e) { return null; }
-    };
-
-    // A collision with one of the roots already there would be remarkable, but stepping the suffix
-    // is what OpenSSL itself does, and the log has to name the file it really wrote.
-    const place = (hash) => {
-        const slots = Array.from({ length: SLOTS }, (_, suffix) => ({
-            suffix,
-            file: path.join(certs, `${hash}.${suffix}`)
-        }));
-
-        const already = slots.find((slot) => contentsOf(slot.file) === ca.cert);
-        if (already) return `${hash}.${already.suffix} already there`;
-
-        const free = slots.find((slot) => contentsOf(slot.file) === null);
-        if (!free) return `${hash} has no free slot`;
-
-        fs.writeFileSync(free.file, ca.cert);
-        return `${hash}.${free.suffix} written`;
-    };
-
-    const hashes = x509.subjectHashes(ca.commonName);
-
-    note('trusted', `${ca.commonName} — ${[hashes.hash, hashes.hashOld].map(place).join(', ')}`);
-};
-
-const cobaltIsInstalledHere = () => {
-    try { return fs.statSync(STOCK).isDirectory(); } catch (e) { return false; }
-};
-
-const stagedMaterial = (content) => {
-    const copied = copyInto(STOCK, content);
-    if (copied) note('staged', `${copied} files into ${content}`);
-
-    return existingMaterial();
-};
-
-const stageOrFail = (content) => {
-    try {
-        return { material: stagedMaterial(content) };
-    } catch (error) {
-        return { error };
-    }
 };
 
 const LAUNCH_AFTER_KILL = 1200;
